pyproject.toml
[project]
name = "adstastic-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
]
